+++ packages/spfx/src/getListItem/index.ts
@@ -39,9 +39,8 @@ async function getListItem<T = Record<string, any>>(options: GetListItemByIDOpti
     const select = getStringArray(options.select);
     const expand = getStringArray(options.expand);
 
-    if (expand.length === 0) {
-      expand.push("AttachmentFiles");
-    }
+    if (!expand.includes("EncodedAbsUrl")) expand.push("EncodedAbsUrl");
+    if (!expand.includes("AttachmentFiles")) expand.push("AttachmentFiles");
 
     const data = await item.select(...select).expand(...expand)();